requirements.txt
openai>=1.30.0
requests>=2.31.0
moviepy>=1.0.3
gTTS>=2.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
streamlit>=1.35.0
imageio-ffmpeg>=0.4.9
numpy>=1.24.0
